// src/modules/system/enums/messages/pagination-messages/pagination-messages.enum.ts
export enum PaginationMessage {
  PAGE_INT = 'Page must be positive integer',
  PAGE_REQUIRED = 'Page is required',
  PAGE_MIN = 'Page should be greater than or equal to 1',

  PAGE_SIZE_INT = 'Page size must be positive integer',
  PAGE_SIZE_REQUIRED = 'Page size is required',
  PAGE_SIZE_MIN = 'Page size should be greater than or equal to 1',
}